Updated genesis protocol editor
